Refine Swift scheduler timestep handling for Euler parity
Move the Stable Diffusion and SDXL scheduler path onto float timesteps so scheduler-emitted values can flow through scheduler scaling, ControlNet, and UNet conditioning without premature integer rounding.

Update the Euler discrete scheduler to preserve fractional timesteps and interpolate sigma values for scaleModelInput, while adapting the existing scheduler implementations to the new float-based protocol surface.

Add fractional-timestep unit coverage and record the current Phase 2 status in PLAN.md. Validated with an Apple-toolchain build and package test pass through Xcode MCP.

+++ swift/StableDiffusion/pipeline/ControlNet.swift
@@ -63,12 +63,12 @@ public struct ControlNet: ResourceManaging {
     /// - Returns: Array of predicted noise residuals
     func execute(
         latents: [MLShapedArray<Float32>],
-        timeStep: Int,
+        timeStep: Float,
         hiddenStates: MLShapedArray<Float32>,
         images: [MLShapedArray<Float32>]
     ) throws -> [[String: MLShapedArray<Float32>]] {
         // Match time step batch dimension to the model / latent samples
-        let t = MLShapedArray(scalars: [Float(timeStep), Float(timeStep)], shape: [2])
+        let t = MLShapedArray(scalars: [timeStep, timeStep], shape: [2])
         
         var outputs: [[String: MLShapedArray<Float32>]] = []
